+++ cosmwasm/src/msg.rs
@@ -20,7 +20,7 @@ pub enum QueryMsg {
     // GetCount returns the current count as a json-encoded number
     GetCount {},
     // GetRequest returns the request ad index.
-    GetRequest { index: usize },
+    GetRequest { index: i32 },
 }
 
 // We define a custom struct for each query response
@@ -32,6 +32,6 @@ pub struct CountResponse {
 // We define a custom struct for each query response
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct RequestResponse {
-    pub id: usize,
+    pub id: i32,
     pub body: String,
 }
